+++ NEWS.md
@@ -1,3 +1,7 @@
+Version 0.11.4 For Dev-C++ 7 Beta
+ - fix: compiler set's custom link parameters  not used when compiling
+ - fix: 
+
 Version 0.11.3 For Dev-C++ 7 Beta
  - fix: use pixel size for fonts, to fit different dpi in multiple displays
  - enhancement: use the new expression parser to parse info for tips